inventory123/backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.2.0
django-filter==23.5
drf-yasg==1.21.7
django-ratelimit==4.1.0
django-axes==6.3.0
django-security==0.20.0
cryptography==42.0.5
django-ssl-auth==0.1.0
django-secure==1.0.2
django-honeypot==1.0.2
django-ipware==6.0.0
django-user-agents==0.4.0
django-request-logging==0.7.0
django-auditlog==2.4.0
django-session-timeout==0.1.0
django-password-validators==0.4.0
django-oauth-toolkit==2.3.0
django-allauth==0.60.1
django-two-factor-auth==1.15.5
django-otp==1.3.0
django-ratelimit-backend==0.1.0
django-defender==0.9.0
django-lockout==0.1.0
django-security-headers==0.1.0
django-csp==3.7
django-xss-protection==0.1.0 